fix: preserve resume dependency boundaries
